pyproject.toml
[tool.poetry]
name = "iracing-yn360"
version = "1.1.2"
description = "Simple application that connects the iRacing flags to a YN360 RGB Light"
authors = ["Chris Champeau <semaj4712@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "iracing_yn360"}]

[tool.poetry.dependencies]
python = "^3.11,<3.12"
bleak = "^0.19.5"
pyirsdk = "^1.3.2"
appdirs = "^1.4.4"
pytimedinput = "^2.0.1"
semantic-version = "^2.10.0"
packaging = "^23.0"
requests = "^2.28.2"
pyinstaller = "^5.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
